Move shared types into components, export them
