+++ packages/argo-admin/src/index.ts
@@ -2,4 +2,3 @@ export * from './api';
 export * from './components';
 export * from './extension-api';
 export * from './extension-points';
-export * from './types';
